Reject nested author fields if last name is blank
Professor Stolley's explanation:

This uses the `reject_if:` option on `accepts_nested_attributes_for`,
and references a custom `incomplete_author` method that checks to see
whether the `last_name` attribute on any nested author fields is blank.

According to the docs linked to in the code comment in this commit, it
is possible to pass `:all_blank` to `reject_if:`, but for author records
one assumes that a corporate author might have only the last name filled
in, so it's better to write a custom method. And of course as author
records get more complex, it would be easy to add additional checks to
to the `incomplete_author` method.